# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Go API Scaffold
  version: 0.1.0
  description: 本仓库提供的通用 Go API 脚手架（认证/用户资料/示例 Todo 模块）。
servers:
  - url: http://localhost:8080
tags:
  - name: system
  - name: auth
  - name: user
  - name: admin

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UnifiedResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
          description: 成功时为业务数据；失败时为 null
          oneOf:
            - type: object
            - type: array
            - type: 'null'
        request_id:
          type: string
          nullable: true

    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64

    LoginResult:
      type: object
      properties:
        tokens:
          $ref: '#/components/schemas/TokenPair'
        user:
          $ref: '#/components/schemas/UserInfo'

    UserInfo:
      type: object
      properties:
        id:
          type: string
        nickname:
          type: string
        avatar:
          type: string
        role:
          type: string

    UserProfile:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        nickname:
          type: string
        avatar:
          type: string
        role:
          type: string

    TodoItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        done:
          type: boolean
        createdAt:
          type: string
          example: "2026-06-05T10:00:00Z"
        updatedAt:
          type: string
          example: "2026-06-05T10:00:00Z"

    TodoListData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodoItem'
        page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64

    RegisterReq:
      type: object
      required: [grant_type]
      properties:
        grant_type:
          type: string
          enum: [password, code]
        code_type:
          type: string
          enum: [sms, email]
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true

    LoginReq:
      type: object
      required: [grant_type]
      properties:
        grant_type:
          type: string
          enum: [password, sms_code, email_code, oauth]
        account:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        provider:
          type: string
          enum: [wechat, apple, google]
          nullable: true
        open_id:
          type: string
          nullable: true
          description: OAuth 开发模式（oauth.dev_mode=true）下直接传第三方 open_id

    SendCodeReq:
      type: object
      required: [type, target]
      properties:
        type:
          type: string
          enum: [sms, email]
        target:
          type: string

    RefreshReq:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    LogoutReq:
      type: object
      properties:
        refresh_token:
          type: string

    UpdateProfileReq:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 30
          nullable: true
        avatar:
          type: string
          maxLength: 500
          nullable: true

    CreateTodoReq:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200

    UpdateTodoReq:
      type: object
      required: [title, done]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        done:
          type: boolean

paths:
  /health:
    get:
      tags: [system]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'

  /ready:
    get:
      tags: [system]
      responses:
        '200':
          description: Ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'

  /api/v1/auth/register:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResult'

  /api/v1/auth/login:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResult'

  /api/v1/auth/send-code:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'

  /api/v1/auth/refresh:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenPair'

  /api/v1/auth/logout:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'

  /api/v1/user/profile:
    get:
      tags: [user]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
    patch:
      tags: [user]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'

  /api/v1/user/todos:
    get:
      tags: [user]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TodoListData'
    post:
      tags: [user]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TodoItem'

  /api/v1/user/todos/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    get:
      tags: [user]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TodoItem'
    patch:
      tags: [user]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnifiedResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TodoItem'
    delete:
      tags: [user]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'

  /api/v1/admin/ping:
    get:
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'